requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
Telethon==1.24.0
TgCrypto==1.2.5
urllib3==2.3.0
